+++ ParallaxGenLib/src/BethesdaGame.cpp
@@ -1,8 +1,19 @@
 #include "BethesdaGame.hpp"
 
-#include <knownfolders.h>
-#include <shlobj.h>
 #include <spdlog/spdlog.h>
+
+#include <guiddef.h>
+#include <shlobj.h>
+
+#include <objbase.h>
+#include <windows.h>
+
+#include <filesystem>
+#include <map>
+#include <stdexcept>
+#include <string>
+
+#include <cstdlib>
 
 using namespace std;
 
@@ -211,44 +222,47 @@ auto BethesdaGame::getGamePath() const -> filesystem::path {
   return GamePath;
 }
 
+auto BethesdaGame::getGameRegistryPath() const -> std::string {
+  const string BasePathSkyrim = R"(SOFTWARE\WOW6432Node\bethesda softworks\)";
+  const string BasePathEnderal = R"(Software\Sure AI\)";
+
+  const std::map<GameType, string> GameToPathMap{{GameType::SKYRIM_SE, BasePathSkyrim + "Skyrim Special Edition"},
+                                                 {GameType::SKYRIM, BasePathSkyrim + "Skyrim"},
+                                                 {GameType::SKYRIM_VR, BasePathSkyrim + "Skyrim VR"},
+                                                 {GameType::ENDERAL, BasePathEnderal + "Enderal"},
+                                                 {GameType::ENDERAL_SE, BasePathEnderal + "EnderalSE"}};
+
+  if (GameToPathMap.contains(ObjGameType)) {
+    return GameToPathMap.at(ObjGameType);
+  }
+
+  return {};
+}
+
 auto BethesdaGame::getGameDataPath() const -> filesystem::path { return GameDataPath; }
 
 auto BethesdaGame::findGamePathFromSteam() const -> filesystem::path {
   // Find the game path from the registry
   // If the game is not found, return an empty string
 
-  // Check if key doesn't exists in steam map
-  if (getSteamGameID() == 0) {
-    return {};
-  }
+  HKEY BaseHKey = (ObjGameType == GameType::ENDERAL || ObjGameType == GameType::ENDERAL_SE) ? HKEY_CURRENT_USER
+                                                                                            : HKEY_LOCAL_MACHINE;
+  const string RegPath = getGameRegistryPath();
 
-  string RegPath =
-      R"(Software\Microsoft\Windows\CurrentVersion\Uninstall\Steam App )" + to_string(this->getSteamGameID());
+  char Data[REG_BUFFER_SIZE]{};
+  DWORD DataSize = REG_BUFFER_SIZE;
 
-  HKEY HKey = nullptr;
-
-  char Data[REG_BUFFER_SIZE]{}; // NOLINT
-  DWORD DataType = REG_SZ;
-  DWORD DataSize = sizeof(Data);
-
-  LONG Result = RegOpenKeyExA(HKEY_LOCAL_MACHINE, RegPath.c_str(), 0, KEY_READ, &HKey);
+  LONG Result = RegGetValueA(BaseHKey, RegPath.c_str(), "Installed Path", RRF_RT_REG_SZ, nullptr, Data, &DataSize);
   if (Result == ERROR_SUCCESS) {
-    // Query the value
-    Result = RegQueryValueExA(HKey, "InstallLocation", nullptr, &DataType, (LPBYTE)Data, &DataSize); // NOLINT
-    if (Result == ERROR_SUCCESS) {
-      // Ensure null-terminated string
-      Data[DataSize] = '\0'; // NOLINT
-      return string(Data);   // NOLINT
-    }
-    // Close the registry key
-    RegCloseKey(HKey);
+    return {Data};
   }
 
   return {};
 }
 
 auto BethesdaGame::getINIPaths() const -> BethesdaGame::ININame {
-  static BethesdaGame::ININame Output = getINILocations();
+  BethesdaGame::ININame Output = getINILocations();
+  const filesystem::path GameDocsPath = getGameDocumentPath();
 
   // normal ini file
   Output.INI = GameDocumentPath / Output.INI;
@@ -285,7 +299,7 @@ auto BethesdaGame::getGameAppdataPath() const -> filesystem::path {
 
 auto BethesdaGame::getSystemPath(const GUID &FolderID) -> filesystem::path {
   PWSTR Path = nullptr;
-  HRESULT Result = SHGetKnownFolderPath(FolderID, 0, nullptr, &Path);
+  const HRESULT Result = SHGetKnownFolderPath(FolderID, 0, nullptr, &Path);
   if (SUCCEEDED(Result)) {
     wstring OutPath(Path);
     CoTaskMemFree(Path); // Free the memory allocated for the path