+++ ParallaxGenLib/src/patchers/PatcherComplexMaterial.cpp
@@ -14,8 +14,9 @@ using namespace ParallaxGenUtil;
 
 // Statics
 std::unordered_set<wstring> PatcherComplexMaterial::DynCubemapBlocklist; // NOLINT
-bool PatcherComplexMaterial::DisableMLP;
+
 ParallaxGenDirectory *PatcherComplexMaterial::PGD;
+bool PatcherComplexMaterial::DisableMLP; // NOLINT
 
 auto PatcherComplexMaterial::loadStatics(const unordered_set<wstring> &DynCubemapBlocklist, const bool &DisableMLP,
                                          ParallaxGenDirectory *PGD) -> void {
@@ -153,7 +154,7 @@ auto PatcherComplexMaterial::applyPatch(NiShape *NIFShape, const wstring &Matche
   NIFUtil::clearShaderFlag(NIFShaderBSLSP, SLSF1_PARALLAX, NIFModified);
   NIFUtil::setShaderFlag(NIFShaderBSLSP, SLSF1_ENVIRONMENT_MAPPING, NIFModified);
   // Set complex material texture
-  string NewParallax;
+  const string NewParallax;
   NIFUtil::setTextureSlot(NIF, NIFShape, NIFUtil::TextureSlots::PARALLAX, NewParallax, NIFModified);
   NIFUtil::setTextureSlot(NIF, NIFShape, NIFUtil::TextureSlots::ENVMASK, MatchedPath, NIFModified);