+++ ParallaxGenLib/src/patchers/PatcherTruePBR.cpp
@@ -227,7 +227,7 @@ auto PatcherTruePBR::getPathContainsMatch(const wstring &LogPrefix,
     }
   }
 
-  wstring SlotLabel = L"path_contains";
+  const wstring SlotLabel = L"path_contains";
   if (NumMatches > 0) {
     spdlog::trace(L"{}PBR | {} | Matched {} PBR JSONs for \"{}\"", LogPrefix, SlotLabel, NumMatches, Diffuse);
   } else {
@@ -320,8 +320,8 @@ auto PatcherTruePBR::applyPatch(NiShape *NIFShape, nlohmann::json &TruePBRData, 
   // Prep
   auto *NIFShader = NIF->GetShader(NIFShape);
   auto *const NIFShaderBSLSP = dynamic_cast<BSLightingShaderProperty *>(NIFShader);
-  bool EnableTruePBR = !MatchedPath.empty();
-  bool EnableEnvMapping = TruePBRData.contains("env_mapping") && TruePBRData["env_mapping"] && !EnableTruePBR;
+  const bool EnableTruePBR = !MatchedPath.empty();
+  const bool EnableEnvMapping = TruePBRData.contains("env_mapping") && TruePBRData["env_mapping"] && !EnableTruePBR;
 
   // "delete" attribute
   if (TruePBRData.contains("delete") && TruePBRData["delete"]) {
@@ -588,7 +588,7 @@ auto PatcherTruePBR::enableTruePBROnShape(NiShape *NIFShape, NiShader *NIFShader
     NIFUtil::setTextureSlot(NIF, NIFShape, NIFUtil::TextureSlots::PARALLAX, NewParallax, NIFModified);
   }
 
-  string Slot4;
+  const string Slot4;
   NIFUtil::setTextureSlot(NIF, NIFShape, NIFUtil::TextureSlots::CUBEMAP, Slot4, NIFModified);
 
   // "lock_rmaos" attribute