+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fixed PBR rename not functions with certain suffixes
 - Fixed ParallaxGen thinking RMAOS files were CM
 - -vv logging mode will only log to file now not the terminal buffer
+- Added a function to classify vanilla BSAs in the config, which will ignore complex material and parallax files from them
+- Fixed retrieval of game installation directories from the registry
 
 ## [0.5.8] - 2025-09-21